pkg: Move promquery package contents into prestostore package
Only prestostore needs this and the callback approach for making the
promquery stuff more generic was a pain to use
